chore: move the requesters to a subfolder
